refactored and added context. MVP reached!
