+++ smurfs/src/contexts/SmurfContext.js
@@ -1,0 +1,2 @@
+import { createContext } from 'react';
+export const SmurfContext = createContext();
